feat: improve error handling and validation in KeyResultsService and ObjectiveService

+++ backend/src/key-results/key-results.service.ts
@@ -2,13 +2,18 @@ import { CreateKeyResultDto } from './dto/create-key-result.dto';
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { UpdateKeyResultDto } from './dto/update-key-result.dto';
+import { KeyResultNotFoundException } from '../common/exceptions/key-result-not-found.exception';
 
 @Injectable()
 export class KeyResultsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  getAll() {
-    return this.prisma.keyResult.findMany();
+  async getAll() {
+    const keyResults = await this.prisma.keyResult.findMany();
+    if (!keyResults) {
+      throw new Error('No key results found');
+    }
+    return keyResults;
   }
 
   async getKeyResultById(id: number) {
@@ -16,7 +21,7 @@ export class KeyResultsService {
       where: { id },
     });
     if (!keyResult) {
-      throw new Error(`Key Result with id ${id} not found`);
+      throw new KeyResultNotFoundException(id);
     }
     return keyResult;
   }
@@ -28,12 +33,20 @@ export class KeyResultsService {
 
     if (!keyResults) {
       throw new Error(
-        `No Key Results found for Objective with id ${objectiveId}`,
+        `No key results found for objective with id ${objectiveId}`,
       );
     }
     return keyResults;
   }
   async create(objectiveId: number, dtos: CreateKeyResultDto[]) {
+    const objective = await this.prisma.objective.findUnique({
+      where: { id: objectiveId },
+    });
+
+    if (!objective) {
+      throw new Error(`Objective with id ${objectiveId} not found`);
+    }
+
     return this.prisma.keyResult.createMany({
       data: dtos.map((dto) => ({
         description: dto.description,
@@ -44,6 +57,13 @@ export class KeyResultsService {
   }
 
   async update(id: number, dto: UpdateKeyResultDto) {
+    const keyResult = await this.prisma.keyResult.findUnique({
+      where: { id },
+    });
+    if (!keyResult) {
+      throw new KeyResultNotFoundException(id);
+    }
+
     return this.prisma.keyResult.update({
       where: { id },
       data: {
@@ -53,7 +73,13 @@ export class KeyResultsService {
     });
   }
 
-  delete(id: number) {
+  async delete(id: number) {
+    const keyResult = await this.prisma.keyResult.findUnique({
+      where: { id },
+    });
+    if (!keyResult) {
+      throw new KeyResultNotFoundException(id);
+    }
     return this.prisma.keyResult.delete({
       where: { id },
     });
